Flatten useless module hierarchy
